+++ test/test_project/lib/data/repo.dart
@@ -1,0 +1,4 @@
+// expect_lint: linesman
+import 'package:linesman_example/ui/widget.dart';
+
+const repo = widget;